+++ test/components/datagrid/datagrid-settings.func-spec.js
@@ -72,6 +72,7 @@ describe('Datagrid Settings', () => { //eslint-disable-line
       onBeforeSelect: null,
       selectChildren: true,
       allowSelectAcrossPages: null,
+      selectAllOnlyActivePage: false,
       groupable: null,
       showNewRowIndicator: true,
       stretchColumn: null,