+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### v4.35.0 Features
 
+- `[Datagrid]` Added support to select all rows on active page only for client side paging. ([#4265](https://github.com/infor-design/enterprise/issues/4265))
 - `[Lookup]` Added the ability to change the lookup icon. ([#4527](https://github.com/infor-design/enterprise/issues/4527))
 
 ### v4.35.0 Fixes